Adjusted name of dataset.
