--- src/pages/apis/tobacco/event/_meta.yaml
@@ -1,13 +1,0 @@
-documentTitle: openFDA › Tobacco › Adverse Events API
-name: Tobacco Adverse Events
-description: "Reports about tobacco products that are damaged, defective, contaminated, smell or taste wrong, or cause undesirable health effects."
-datasets:
-  - tobaccoevent
-label: Tobacco Adverse Events
-title: Tobacco Adverse Events
-path: /apis/tobacco/event
-api_path: /tobacco/event
-start: 20170101
-status: tobaccoevent
-dateConstraintKey: effective_time
-type: endpoint
