+++ src/pages/apis/tobacco/productproblem/example-api-queries/index.jsx
@@ -12,7 +12,7 @@ class IndexRoute extends React.Component {
 
     return (
       <section className="doc-content">
-        <h2>Example tobacco adverse events API queries</h2>
+        <h2>Example tobacco product problem reports API queries</h2>
         <p>To help get you started, we have provided some API query examples below. Use the Run query button to call the API and get back results. You can experiment by editing the example queries in the black text box.</p>
 
         <QueryExplorer
