+++ src/pages/downloads/index.jsx
@@ -23,7 +23,7 @@ import device_recall_meta from '../apis/device/recall/_meta.yaml'
 import device_registration_listing_meta from '../apis/device/registrationlisting/_meta.yaml'
 import device_udi_meta from '../apis/device/udi/_meta.yaml'
 import device_covid19serology_meta from '../apis/device/covid19serology/_meta.yaml'
-import tobacco_event_meta from '../apis/tobacco/event/_meta.yaml'
+import tobacco_productproblem_meta from '../apis/tobacco/productproblem/_meta.yaml'
 import other_nsde_meta from '../apis/other/nsde/_meta.yaml'
 import other_substance_meta from '../apis/other/substance/_meta.yaml'
 
@@ -51,7 +51,7 @@ var endpoint_list = {
   'Medical Device UDI': device_udi_meta,
   'COVID-19 Serological Testing Evaluations': device_covid19serology_meta,
   'Tobacco': 'tobacco_header',
-  'Tobacco Event': tobacco_event_meta,
+  'Tobacco Product Problem': tobacco_productproblem_meta,
   'Other': 'other_header',
   'NSDE': other_nsde_meta,
   'Substance': other_substance_meta
